+++ src/Telegram.php
@@ -457,6 +457,13 @@ class Telegram
         //This is necessary to "require" all the necessary command files!
         $this->getCommandsList();
 
+        //Make sure we don't try to process update that was already processed
+        $last_id = DB::selectTelegramUpdate(1, $this->update->getUpdateId());
+        if ($last_id && count($last_id) === 1) {
+            TelegramLog::debug('Duplicate update received, processing aborted!');
+            return Request::emptyResponse();
+        }
+
         DB::insertRequest($this->update);
 
         return $this->executeCommand($command);